-Added a simple Abs BmodeFilter -Made a correction for CPU USImage Beamforming

+++ Modules/PhotoacousticsAlgorithms/Algorithms/OCL/mitkPhotoacousticOCLBeamformer.h
@@ -104,7 +104,7 @@ protected:
 
   int GetBytesPerElem()
   {
-    return sizeof(double);
+    return sizeof(float);
   }
 
   virtual us::Module* GetModule();
